+++ modelforge/dataset/utils.py
@@ -1,0 +1,1 @@
+# generates and retrieves hdf5 files from zenodo, defines the interaction with the hdf5 file
